This is the initial version of the codebot released to the public.
Basically this is a wrapper around native programming enviroment tools such as ruby/irb/python/python console/g++/gcc/javac/java/compiled code. It will run any or all
of these as child processes and communicate with them through stdin/stdout/stderr. Codebot will parse this info and makes it available through irc.
To be able to run it you need ruby 1.9.2 and the open4 ruby gem/lib. To make use of all the functionality the above programs should be available on the path.

The basic code structure is the following:
  - All low lever irc client stuff is in IrcClient.rb. It keeps the irc client connected to the channel and maintains a userlist
  - CodeBot.rb inherits from IrcClient.rb and communicates with irc by overriding receive/receive_from/receive_private/receive_private from and calling send/send_to_user.
  - CodeBot maintains a map {user -> programming enviroment}. So a different programming enviroment for every user unless explicitly shared. All commands execept new/and help are routed through the programming enviroments
  - Enviroment

Security considerations:

By exposing an unrestricted programming enviroment, you are exposing all the priviliges of a regular user to the people on irc. So in order to make
it secure, you have to restrict these use privileges so no real harm can done to your system. Think of it as a guest shell account. Limitations should include in order of importance:

- No write access to critical files, including in the home directory of the user running the bot
- No read access to user files
- Limited number of threads, otherwise fork calls will kill your system
- Disk quota
- memory and cpu quota

Programs are ristricted to a running time of 5 seconds, after that it will kill the process and all child processes it spawned, but bad things can happen in 5 seconds